Find manifest for download by tags instead of commits
This makes it possible to use `./wpt manifest` on a checkout that is
more than 50 commits behind master. The downside is that it will require
more network requests if the most recent tag/tags don't have releases
uploaded.

Add more warnings to make it clearer why it goes wrong when it does.