Worker: Separate worker-from-blob-url.window.html to filter out failing tests in a more fine-grained way
This test file should be separated into 2 files (one for dedicated workers and
the other for shared workers). This is because Edge and Safari don't support
shared workers, and Chrome has an issue on shared workers with Blob URL
revocation, so some tests are expected to fail on them. By separating the file,
browsers can filter out failing tests in a more fine grained way.

Bug: 655458, 800898
Change-Id: I9ea6c76dc9b6841bf6aaa98251c81f292dd460cf